-CCModel class, move image_info funcs to from_fits module
